cffi: bytes, part3: opportunistic str
This commit partially reverts what part1 did, but hey, that's how the
thinking process went, and now we have it on the record.

While "always return bytes" sounds like a nice and consistent statement,
in reality it makes for a more cumbersome API.

So, here is what we do now: both str and bytes are accepted, as long as
str can be encoded with ASCII encoding. Likewise, when parsing any
result, if a string field contains an ASCII encoded string we return
str, otherwise we return bytes.

Previusly (before all this refactor) pycares did utf-8 encoding, which
is not right, since DNS doesn't specify that.